fix: 修复PageResponse类型转换问题

+++ backend/src/main/java/com/practice/service/QuestionService.java
@@ -41,8 +41,8 @@ public class QuestionService {
         
         PageResponse<Question> response = new PageResponse<>();
         response.setTotal(result.getTotal());
-        response.setPage(result.getCurrent());
-        response.setSize(result.getSize());
+        response.setPage(result.getCurrent().intValue());
+        response.setSize(result.getSize().intValue());
         response.setRecords(result.getRecords());
         
         return response;
